+++ GameplanAPI/Features/Season/DeleteSeason/DeleteSeasonCommandHandler.cs
@@ -1,12 +1,14 @@
 ﻿using GameplanAPI.Common.Errors;
 using GameplanAPI.Common.Interfaces;
 using GameplanAPI.Common.Models;
+using GameplanAPI.Common.Services._Interfaces;
 using GameplanAPI.Features.Season._Interfaces;
 
 namespace GameplanAPI.Features.Season.DeleteSeason
 {
     public sealed class DeleteSeasonCommandHandler(
-        ISeasonRepository seasonRepository, 
+        ISeasonRepository seasonRepository,
+        IAuthService authService,
         IUnitOfWork unitOfWork)
         : ICommandHandler<DeleteSeasonCommand>
     {
@@ -19,6 +21,13 @@ namespace GameplanAPI.Features.Season.DeleteSeason
             if (season == null)
             {
                 return Result.Failure(Errors<Season>.NotFound(request.Id));
+            }
+
+            var currentUserId = authService.GetCurrentUserId();
+
+            if (currentUserId == null || season.UserId != currentUserId)
+            {
+                return Result.Failure(Errors<Season>.NotAuthorized);
             }
 
             seasonRepository.Delete(season);
